+++ apps/nextjs-website/react-components/components/Header/Header.tsx
@@ -1,304 +1,15 @@
-import {
-  Box,
-  Link,
-  Stack,
-  Typography,
-  Divider,
-  useTheme,
-  Theme,
-  Chip,
-  List,
-  ListItem,
-  ListItemText,
-} from '@mui/material';
-import CloseIcon from '@mui/icons-material/Close';
-import MenuIcon from '@mui/icons-material/Menu';
+import React from 'react';
+import { Box, Link, Stack, Typography, Divider } from '@mui/material';
 import ChatBubbleOutlineIcon from '@mui/icons-material/ChatBubbleOutline';
-import { ArrowDropDown } from '@mui/icons-material';
 import { useMediaQuery, useTheme as useMuiTheme } from '@mui/material';
-import {
-  HeaderProps,
-  MenuDropdownProp,
-  DropdownItem,
-  HeaderTitleProps,
-  NavigationProps,
-  DialogBubbleProps,
-} from '@react-components/types/Header/Header.types';
-import {
-  BackgroundColor,
-  TextAlternativeColor,
-  TextColor,
-} from '@react-components/components/common/Common.helpers';
+import { HeaderProps } from '@react-components/types/Header/Header.types';
+import { BackgroundColor } from '@react-components/components/common/Common.helpers';
 import { CtaButtons } from '../common/Common';
 import { useState, useEffect } from 'react';
-
-const useStyles = ({ theme, active }: MenuDropdownProp, { spacing }: Theme) => {
-  const textColor = TextAlternativeColor(theme);
-
-  return {
-    menu: {
-      borderColor: textColor,
-      borderBottomStyle: 'solid',
-      width: '100%',
-      borderBottomWidth: 3,
-      borderBottomColor: {
-        md: active ? textColor : 'transparent',
-        xs: 'transparent',
-      },
-      backgroundColor: { xs: 'white', md: 'transparent' },
-      height: '100%',
-    },
-    link: {
-      textIndent: { xs: spacing(2), md: 0 },
-      display: 'flex',
-      alignItems: 'center',
-      height: '100%',
-      padding: '0 8px',
-    },
-    arrowAnimate: {
-      transition: 'transform 0.2s',
-    },
-  };
-};
-
-const DialogBubble = ({ children, ...stackProps }: DialogBubbleProps) => {
-  const muiTheme = useTheme();
-  const styles = {
-    bubbleContainer: {
-      position: 'absolute',
-      marginTop: muiTheme.spacing(8),
-      padding: muiTheme.spacing(2),
-      direction: 'ltr',
-      textAlign: { xs: 'right', md: 'left' },
-      borderRadius: '4px',
-      zIndex: 1,
-    },
-  };
-
-  return (
-    <Stack
-      sx={{
-        ...styles.bubbleContainer,
-        bgcolor: 'common.white',
-        boxShadow: { xs: 'custom.boxShadow', md: 'custom.otMenuMobile' },
-      }}
-      aria-haspopup='true'
-      {...stackProps}
-    >
-      {children}
-    </Stack>
-  );
-};
-
-const HamburgerMenu = ({
-  open,
-  onOpen,
-  onClose,
-}: {
-  open: boolean;
-  onOpen: () => void;
-  onClose: () => void;
-}) => {
-  const muiTheme = useTheme();
-
-  return open ? (
-    <CloseIcon
-      sx={{ color: muiTheme.palette.text.secondary }}
-      cursor='pointer'
-      onClick={onClose}
-      aria-label='chiudi'
-      aria-haspopup='true'
-      aria-expanded='true'
-    />
-  ) : (
-    <MenuIcon
-      sx={{ color: muiTheme.palette.text.secondary }}
-      cursor='pointer'
-      onClick={onOpen}
-      aria-label='apri'
-      aria-haspopup='true'
-      aria-expanded='false'
-    />
-  );
-};
-
-const MenuDropdown = (props: MenuDropdownProp) => {
-  const { label, active, theme, items, isOpen, onClick, isMobile, ...button } = props;
-  const muiTheme = useTheme();
-  const styles = useStyles(props, muiTheme);
-  const hasLinks = items?.length;
-
-  return (
-    <Stack sx={styles.menu}>
-      <Box
-        sx={{
-          width: '100%',
-          minWidth: 'max-content',
-          height: '100%',
-          display: 'flex',
-          flexDirection: 'row',
-          justifyContent: { xs: 'left', md: 'center' },
-          alignItems: 'center',
-        }}
-      >
-        <Link
-          sx={styles.link}
-          style={{
-            color: active
-              ? muiTheme.palette.primary.main
-              : muiTheme.palette.text.secondary,
-            textDecoration: 'none',
-          }}
-          {...button}
-          onClick={onClick}
-        >
-          <Typography
-            variant='sidenav'
-            color='inherit'
-            sx={{
-              display: 'flex',
-              textDecoration: 'none',
-              fontSize: '1em',
-              padding: 0,
-            }}
-          >
-            {label}
-            {hasLinks && (
-              <ArrowDropDown
-                color='inherit'
-                fontSize='small'
-                sx={{
-                  transition: 'transform 0.2s',
-                  transform: isOpen ? 'rotate(180deg)' : 'rotate(0deg)',
-                  marginLeft: '4px',
-                  cursor: 'pointer',
-                }}
-              />
-            )}
-          </Typography>
-        </Link>
-      </Box>
-      {hasLinks && isOpen && (
-        isMobile ? (
-          <List component="div" disablePadding>
-            {items?.map((item: DropdownItem, index) => (
-              <ListItem button key={item.key ?? index} sx={{ pl: 4 }}>
-                <ListItemText primary={item.label} primaryTypographyProps={{ style: { color: muiTheme.palette.text.secondary } }} />
-              </ListItem>
-            ))}
-          </List>
-        ) : (
-          <DialogBubble>
-            <Stack gap={1}>
-              {items?.map((item: DropdownItem, index) => (
-                <Link
-                  variant='body1'
-                  underline='none'
-                  key={item.key ?? index}
-                  sx={styles.link}
-                  style={{
-                    color: muiTheme.palette.text.secondary,
-                    textDecoration: 'none',
-                    fontSize: '1em',
-                    fontWeight: 600,
-                    padding: 0,
-                  }}
-                  {...item}
-                >
-                  {item.label}
-                </Link>
-              ))}
-            </Stack>
-          </DialogBubble>
-        )
-      )}
-    </Stack>
-  );
-};
-
-const Navigation = ({ menu, theme, isMobile }: NavigationProps & { isMobile: boolean }) => (
-  <Stack
-    gap={{ md: 4, xs: 0 }}
-    direction={{ md: 'row', xs: 'column' }}
-    component='nav'
-    aria-label='main'
-    className='desktop-menu'
-    sx={{
-      width: { xs: '100%', md: 'auto' },
-      height: '100%',
-      alignItems: 'flex-end',
-    }}
-  >
-    {menu.map((menu, index) => (
-      <MenuDropdown
-        key={index}
-        {...menu}
-        theme={theme}
-        isMobile={isMobile}
-        sx={{ py: { xs: 1, sm: 0 }, alignItems: 'flex-end' }}
-      />
-    ))}
-  </Stack>
-);
-
-const HeaderTitle = ({
-  beta,
-  product: { name: productName, href: productHref },
-  theme,
-  logo,
-}: HeaderTitleProps) => {
-  const { spacing } = useTheme();
-  const textColor = TextColor(theme);
-  const label = 'beta';
-
-  return (
-    <Stack
-      direction='row'
-      alignItems='center'
-      gap={1}
-      paddingX={0}
-      paddingY={0}
-    >
-      {logo && (
-        <Link
-          alignItems='center'
-          sx={{
-            display: 'flex',
-            justifyContent: 'center',
-            padding: 0,
-          }}
-          href={logo.href}
-        >
-          <img src={logo.src} alt={logo.alt} style={{ height: 70 }} />
-        </Link>
-      )}
-      {!logo && (
-        <Typography
-          color={textColor}
-          fontSize={{ xs: spacing(3), sm: spacing(3.5) }}
-          noWrap
-          variant='h5'
-        >
-          {productHref ? (
-            <Link href={productHref} underline='none' color='text.primary'>
-              <b>{productName}</b>
-            </Link>
-          ) : (
-            <b>{productName}</b>
-          )}
-        </Typography>
-      )}
-      {beta && (
-        <Chip
-          label={label}
-          color='primary'
-          sx={{ height: 20, width: 45 }}
-          size='small'
-        />
-      )}
-    </Stack>
-  );
-};
+import DesktopDrawer from './components/Header.DesktopDrawer.helpers';
+import { HeaderTitle } from './components/Header.HeaderTitle.helpers';
+import { Navigation } from './components/Header.Navigation.helpers';
+import { HamburgerMenu } from './components/Header.HamburgerMenu.helpers';
 
 const Header = ({
   product,
@@ -324,6 +35,11 @@ const Header = ({
     setMenuOpen(false);
   };
 
+  const [isDrawerOpen, setIsDrawerOpen] = React.useState(false);
+
+  const handleOpenDrawer = () => setIsDrawerOpen(true);
+  const handleCloseDrawer = () => setIsDrawerOpen(false);
+
   useEffect(() => {
     window.addEventListener('resize', closeHeader);
 
@@ -332,9 +48,13 @@ const Header = ({
     };
   }, []);
 
-  const HeaderCtas = () => {
+  interface HeaderCtasProps {
+    onOpenDrawer: () => void;
+  }
+
+  const HeaderCtas: React.FC<HeaderCtasProps> = ({ onOpenDrawer }) => {
     return ctaButtons && ctaButtons.length > 0 ? (
-      <Stack direction='row'>
+      <Stack direction='row' onClick={onOpenDrawer}>
         {CtaButtons({
           ctaButtons: ctaButtons.map((CtaButton) => ({
             ...CtaButton,
@@ -438,7 +158,12 @@ const Header = ({
               isMobile={false}
             />
             <Box sx={{ display: { xs: 'none', md: 'block' } }}>
-              <HeaderCtas />
+              <HeaderCtas onOpenDrawer={handleOpenDrawer} />
+              <DesktopDrawer
+                isOpen={isDrawerOpen}
+                onClose={handleCloseDrawer}
+                anchor='right'
+              />
             </Box>
           </Stack>
         )}
@@ -460,12 +185,16 @@ const Header = ({
                 onClose={closeHeader}
                 open={menuOpen}
               />
-              <Typography variant='body1' color='text.secondary' sx={{ fontWeight: 600 }}>
+              <Typography
+                variant='body1'
+                color='text.secondary'
+                sx={{ fontWeight: 600 }}
+              >
                 Menu
               </Typography>
             </Stack>
             <Box sx={{ display: { xs: 'block', md: 'none' } }}>
-              <HeaderCtas />
+              <HeaderCtas onOpenDrawer={handleOpenDrawer} />
             </Box>
           </Stack>
         )}
